* stats are now cached in memory. this reduce CPU usage of stats collection by 20. Thanks to Franck Villaume

+++ ArchipelServer/modules/hypervisor_health/archipelStatsCollector.py
@@ -30,42 +30,59 @@ class TNThreadedHealthCollector(Thread):
     """
     this class collects hypervisor stats regularly
     """
-    def __init__(self, database_file, collection_interval, max_rows_before_purge): #, snmp_agent, snmp_community, snmp_version, snmp_port):
+    def __init__(self, database_file, collection_interval, max_rows_before_purge, min_row_before_storing): #, snmp_agent, snmp_community, snmp_version, snmp_port):
         """
         the contructor of the class
         """
-        self.database_file = database_file
-        self.collection_interval = collection_interval
-        self.max_rows_before_purge = max_rows_before_purge
-        
-        #self.snmp_community     = cmdgen.CommunityData(snmp_agent, snmp_community, snmp_version)
-        #self.snmp_udptransport  = cmdgen.UdpTransportTarget(('localhost', snmp_port))
-        
-        # ## uname
-        # errorIndication, errorStatus, errorIndex, varBinds = cmdgen.CommandGenerator().getCmd(self.snmp_community, self.snmp_udptransport, (1,3,6,1,2,1,1,1,0))
-        # uname = str(varBinds[0][1])
-        # uname = uname.split()
-        # self.uname_stats = {"krelease": uname[2] , "kname": uname[0] , "machine": uname[1], "os": uname[11]}
+        self.database_file          = database_file
+        self.collection_interval    = collection_interval
+        self.max_rows_before_purge  = max_rows_before_purge
+        self.min_row_before_storing = min_row_before_storing
+        self.stats_CPU              = []
+        self.stats_memory           = []
+        self.stats_load             = []
+        self.stats_disks            = []
         
         uname = commands.getoutput("uname -rsmo").split()
         self.uname_stats = {"krelease": uname[0] , "kname": uname[1] , "machine": uname[2], "os": uname[3]}
         
-        
-        #log.info("opening stats database file {0}".format(self.database_file))
-        
         self.database_query_connection = sqlite3.connect(self.database_file)
-        
-        
-        #self.query_database_connection = sqlite3.connect(self.database_file)
         self.cursor = self.database_query_connection.cursor()
         
         self.cursor.execute("create table if not exists cpu (collection_date date, idle int)")
         self.cursor.execute("create table if not exists memory (collection_date date, free integer, used integer, total integer, swapped integer)")
         self.cursor.execute("create table if not exists disk (collection_date date, total int, used int, free int, free_percentage int)")
         self.cursor.execute("create table if not exists load (collection_date date, one float, five float, fifteen float)")
-        #log.info("Database ready.")
-            
+        log.info("Database ready.")
+        
+        self.recover_stored_stats()
+        
+        
         Thread.__init__(self)
+    
+    
+    def recover_stored_stats(self):
+        log.info("recovering stored statistics")
+        self.cursor.execute("select * from cpu order by collection_date desc limit %d" % self.max_rows_before_purge)
+        for values in self.cursor:
+            date, idle = values
+            self.stats_CPU.insert(0, {"date": date, "id": idle})
+        
+        self.cursor.execute("select * from memory order by collection_date desc limit %d" % self.max_rows_before_purge)
+        for values in self.cursor:
+            date, free, used, total, swapped = values
+            self.stats_memory.insert(0, {"date": date, "free": free, "used" : used, "total": total, "swapped": swapped})
+        
+        self.cursor.execute("select * from load order by collection_date desc limit %d" % self.max_rows_before_purge)
+        for values in self.cursor:
+            date, one, five, fifteen = values
+            self.stats_load.insert(0, {"date": date, "one": one, "five": five, "fifteen": fifteen})
+        
+        self.cursor.execute("select * from disk order by collection_date desc limit %d" % self.max_rows_before_purge)
+        for values in self.cursor:
+            date, total, used, free, free_percentage = values
+            self.stats_disks.insert(0, {"date": date, "total": total, "used": used, "free": free, "free_percentage": free_percentage})
+        log.info("statistics recovered")
     
     
     def get_collected_stats(self, limit=1):
@@ -76,115 +93,48 @@ class TNThreadedHealthCollector(Thread):
         """        
         #log.debug("Retrieving last "+ str(limit) + " recorded stats data for sending")
         try:
-            self.cursor.execute("select * from cpu order by collection_date desc limit " + str(limit))
-            cpu_stats = []
-            for values in self.cursor:
-                date, idle = values
-                cpu_stats.append({"date": date, "id": idle})
-        
-            self.cursor.execute("select * from memory order by collection_date desc limit " + str(limit))
-            memory_stats = []
-            for values in self.cursor:
-                date, free, used, total, swapped = values
-                memory_stats.append({"date": date, "free": free, "used" : used, "total": total, "swapped": swapped})
-        
-            self.cursor.execute("select * from disk order by collection_date desc limit " + str(limit))
-            disk_stats = []
-            for values in self.cursor:
-                date, total, used, free, free_percentage = values
-                disk_stats.append({"date": date, "total": total, "used": used, "free": free, "free_percentage": free_percentage})
-        
-            self.cursor.execute("select * from load order by collection_date desc limit " + str(limit))
-            load_stats = []
-            for values in self.cursor:
-                date, one, five, fifteen = values
-                load_stats.append({"date": date, "one": one, "five": five, "fifteen": fifteen})
-        
-            # uptime
-            # errorIndication, errorStatus, errorIndex, varBinds = cmdgen.CommandGenerator().getCmd(self.snmp_community, self.snmp_udptransport, (1,3,6,1,2,1,1,3,0))
-            # uptime = str(varBinds[0][1])
-            # uptime_stats = {"up" : uptime}
             uptime = commands.getoutput("uptime").split("up ")[1].split(",")[0]
             uptime_stats = {"up" : uptime}
-        
-            return {"cpu": cpu_stats, "memory": memory_stats, "disk": disk_stats, "load": load_stats, "uptime": uptime_stats, "uname": self.uname_stats}
+            return {"cpu": self.stats_CPU[0:limit], "memory": self.stats_memory[0:limit], "disk": self.stats_disks[0:limit], "load": self.stats_load[0:limit], "uptime": uptime_stats, "uname": self.uname_stats}
         except Exception as ex:
             log.error("stat recuperation fails. Exception %s" % str(ex))
             return None
     
     
     def get_memory_stats(self):
-        # errorIndication, errorStatus, errorIndex, varBinds = cmdgen.CommandGenerator().bulkCmd(self.snmp_community, self.snmp_udptransport, 0, 25, (1,3,6,1,4,1,2021,4))
-        # 
-        # for result in varBinds:
-        #     oid, value = result[0]
-        #     if oid == (1,3,6,1,4,1,2021,4,5,0):
-        #         memTotal = int(value)
-        #     elif oid == (1,3,6,1,4,1,2021,4,11,0):
-        #         memFree = int(value)
-        #     elif oid == (1,3,6,1,4,1,2021,4,6,0):
-        #         memUsed = int(value)
-        #     elif oid == (1,3,6,1,4,1,2021,4,3,0):
-        #         swapped = int(value)
-        
         file_meminfo    = open('/proc/meminfo')
         meminfo         = file_meminfo.read()
         file_meminfo.close()
-
+        
         meminfolines    = meminfo.split("\n")
         memTotal        = int(meminfolines[0].split()[1])
         memFree         = int(meminfolines[1].split()[1])
         swapped         = int(meminfolines[4].split()[1])
         memUsed         = memTotal - memFree
         
-        return (datetime.datetime.now(), memFree, memUsed, memTotal, swapped)
+        return {"date": datetime.datetime.now(), "free": memFree, "used" : memUsed, "total": memTotal, "swapped": swapped}
     
     
     def get_cpu_stats(self):
-        # # CPU idle %
-        # errorIndication, errorStatus, errorIndex, varBinds = cmdgen.CommandGenerator().getCmd(self.snmp_community, self.snmp_udptransport, (1,3,6,1,4,1,2021,11,11,0))    
-        # cpuPct = int(varBinds[0][1])
-        
         dt      = self.deltaTime(1)
         cpuPct  = (dt[len(dt) - 1] * 100.00 / sum(dt))
         
-        return (datetime.datetime.now(), cpuPct)
+        return {"date": datetime.datetime.now(), "id": cpuPct}
     
     
     def get_load_stats(self):
-        # errorIndication, errorStatus, errorIndex, varBinds = cmdgen.CommandGenerator().bulkCmd(self.snmp_community, self.snmp_udptransport, 0, 25, (1,3,6,1,4,1,2021,10,1,3))
-        # for result in varBinds:
-        #     oid, value = result[0]
-        #     if oid == (1,3,6,1,4,1,2021,10,1,3,1):
-        #         load1min = float(str(value))
-        #     elif oid == (1,3,6,1,4,1,2021,10,1,3,2):
-        #         load5min = float(str(value))
-        #     elif oid == (1,3,6,1,4,1,2021,10,1,3,3):
-        #         load15min = float(str(value))
-        
         load_average = commands.getoutput("uptime").split("load average:")[1].split(", ")
         load1min, load5min, load15min = (float(load_average[0]), float(load_average[1]), float(load_average[2]))
         
-        return (datetime.datetime.now(), load1min, load5min, load15min)
+        return {"date": datetime.datetime.now(), "one": load1min, "five": load5min, "fifteen": load15min}
     
     
     def get_disk_stats(self):
-        # errorIndication, errorStatus, errorIndex, varBinds = cmdgen.CommandGenerator().bulkCmd(self.snmp_community, self.snmp_udptransport, 0, 25, (1,3,6,1,4,1,2021,9,1))
-        # for result in varBinds:
-        #     oid, value = result[0]
-        #     if oid == (1,3,6,1,4,1,2021,9,1,6,1):
-        #         total = float(str(value))
-        #     elif oid == (1,3,6,1,4,1,2021,9,1,8,1):
-        #         used = float(str(value))
-        #     elif oid == (1,3,6,1,4,1,2021,9,1,7,1):
-        #         free = float(str(value))
-        #     elif oid == (1,3,6,1,4,1,2021,9,1,9,1):
-        #         freePrct = float(str(value))
-        
         disk_free = commands.getoutput("df -h --total | grep total").split()
         total, used, free, freePrct = (disk_free[1], disk_free[2],disk_free[3], disk_free[4])
         
-        return (datetime.datetime.now(), total, used, free, freePrct)
+        return {"date": datetime.datetime.now(), "total" : total, "used": used, "free": free, "free_percentage": freePrct}
+    
     
     def run(self):
         """
@@ -194,21 +144,34 @@ class TNThreadedHealthCollector(Thread):
         
         while(1):
             try:
-                self.database_thread_connection.execute("insert into memory values(?,?,?,?,?)", self.get_memory_stats())
-                self.database_thread_connection.execute("insert into cpu values(?,?)", self.get_cpu_stats())
-                self.database_thread_connection.execute("insert into load values(?,?,?,?)", self.get_load_stats())
-                self.database_thread_connection.execute("insert into disk values(?,?,?,?,?)", self.get_disk_stats())
-            
-                if int(self.database_thread_connection.execute("select count(*) from memory").fetchone()[0]) >= self.max_rows_before_purge:
-                    #log.debug("Purging the last entries.")
-                    self.database_thread_connection.execute("delete from cpu where collection_date=(select collection_date from cpu order by collection_date asc limit "+ str(self.max_rows_before_purge) +")")
-                    self.database_thread_connection.execute("delete from memory where collection_date=(select collection_date from memory order by collection_date asc limit "+ str(self.max_rows_before_purge) +")")
-                    self.database_thread_connection.execute("delete from load where collection_date=(select collection_date from load order by collection_date asc limit "+ str(self.max_rows_before_purge) +")")
-                    self.database_thread_connection.execute("delete from disk where collection_date=(select collection_date from disk order by collection_date asc limit "+ str(self.max_rows_before_purge) +")")
-        
-                self.database_thread_connection.commit()
-            
-                #log.debug("Stats collected")
+                self.stats_CPU.insert(0, self.get_cpu_stats())
+                self.stats_memory.insert(0, self.get_memory_stats())
+                self.stats_load.insert(0, self.get_load_stats())
+                self.stats_disks.insert(0, self.get_disk_stats())
+                
+                if len(self.stats_CPU) % self.min_row_before_storing == 0:
+                    self.database_thread_connection.executemany("insert into memory values(:date, :free, :used, :total, :swapped)", self.stats_memory)
+                    self.database_thread_connection.executemany("insert into cpu values(:date, :id)", self.stats_CPU)
+                    self.database_thread_connection.executemany("insert into load values(:date, :one , :five, :fifteen)", self.stats_load)
+                    self.database_thread_connection.executemany("insert into disk values(:date, :total, :used,:free, :free_percentage)", self.stats_disks)
+                    log.info("stats saved in database file")
+                    
+                    if len(self.stats_CPU) >= self.min_row_before_storing * 2:
+                        del self.stats_CPU[(-self.min_row_before_storing - 1)]
+                        del self.stats_memory[(-self.min_row_before_storing - 1)]
+                        del self.stats_load[(-self.min_row_before_storing - 1)]
+                        del self.stats_disks[(-self.min_row_before_storing - 1)]
+                        log.info("cached stats have been purged from memory")
+                        
+                    if int(self.database_thread_connection.execute("select count(*) from memory").fetchone()[0]) >= self.max_rows_before_purge:
+                        self.database_thread_connection.execute("delete from cpu where collection_date=(select collection_date from cpu order by collection_date asc limit "+ str(self.max_rows_before_purge) +")")
+                        self.database_thread_connection.execute("delete from memory where collection_date=(select collection_date from memory order by collection_date asc limit "+ str(self.max_rows_before_purge) +")")
+                        self.database_thread_connection.execute("delete from load where collection_date=(select collection_date from load order by collection_date asc limit "+ str(self.max_rows_before_purge) +")")
+                        self.database_thread_connection.execute("delete from disk where collection_date=(select collection_date from disk order by collection_date asc limit "+ str(self.max_rows_before_purge) +")")
+                        log.debug("old stored stats have been purged from memory")
+                    
+                    self.database_thread_connection.commit()
+                
                 time.sleep(self.collection_interval)
             except Exception as ex:
                 log.error("stat collection fails. Exception %s" % str(ex))
